use vacuum INDEX_CLEANUP when available
Postgresql 12 and later may skip automatic index cleanup during vacuum if only a small number of rows have been modified. This was a useful change in general. But it can interfere with reducing the size of the table by pgcompacttable if there are dead tuples in the tail of this table, and the vacuum does not remove them because it thinks that too few lines have changed. We will specify INDEX_CLEANUP for each vacuum to force index cleanup.